requirements.txt
Flask>=2.3
python-dotenv>=1.0
psycopg2-binary>=2.9
sentence-transformers>=2.2
torch>=2.0
pandas>=2.0
matplotlib>=3.7
requests>=2.31
pdfplumber>=0.10
python-docx>=1.1
# LLM provider SDKs — install whichever matches LLM_PROVIDER in .env
openai>=1.0
google-genai>=0.3
